requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchainhub
faiss-cpu
pymupdf
python-dotenv
langchain-ollama
langchain-community